pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "esp-remote"
version = "0.1.0"
description = "Remote ESP development with device registry and multi-device support"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
dependencies = [
    "click>=8.1.0",
    "rich>=13.0.0",
    "paramiko>=3.4.0",
    "pyserial>=3.5",
    "tomli>=2.0.0",
    "tomli-w>=1.0.0",
    "gitpython>=3.1.0",
    "esptool>=4.0",
]

[project.scripts]
esp-remote = "esp_remote.cli:main"
